H1203_2

+++ lib/pluto_source_impl.h
@@ -40,8 +40,8 @@ public:
     //                  bool auto_filter = true);
     pluto_source_impl(const std::string &uri,
                          double frequency,
-                         unsigned long samplerate,
-                         unsigned long bandwidth,
+                         double samplerate,
+                         double bandwidth,
                          bool rx1_en, bool rx2_en,
                          unsigned long buffer_size,
                          bool quadrature, bool rfdc, bool bbdc,
@@ -111,8 +111,8 @@ private:
 
     // Runtime configuration parameters
     double                    frequency{0};
-    unsigned long             samplerate{0};
-    unsigned long             bandwidth{0};
+    double                    samplerate{0};
+    double                    bandwidth{0};
     bool                      rx1_en{true};     // I channel
     bool                      rx2_en{true};    // Q channel
     bool                      quadrature{true};
